pyproject.toml
[project]
name = "youngcan-wyckoff-analysis"
version = "0.1.5"
requires-python = ">=3.11"
description = "Wyckoff method quantitative analysis system for A-shares"
readme = "README.md"
license = "AGPL-3.0-only"
authors = [{ name = "youngcan" }]
keywords = ["wyckoff", "stock", "a-share", "quantitative", "agent"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "akshare>=1.18.9",
    "baostock",
    "efinance>=0.2.0",
    "tushare>=1.2.89",
    "pandas>=2.2.0,<3.0",
    "google-genai>=0.6.0",
    "requests>=2.31.0",
    "certifi",
    "tavily-python>=0.3.0",
    "sqlalchemy>=2.0.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.0.0",
    "py-mini-racer>=0.6.0",
    "matplotlib>=3.8.0",
    "pillow>=10.0.0",
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "anthropic>=0.30.0",
    "openai>=1.30.0",
]

[project.optional-dependencies]
streamlit = [
    "streamlit>=1.52.2",
    "streamlit-javascript>=0.1.5",
    "google-adk>=1.0.0",
    "litellm>=1.40.0",
]
dev = ["pytest>=7.0"]

[project.scripts]
wyckoff = "cli.__main__:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["agents*", "app*", "cli*", "core*", "integrations*", "tools*", "utils*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
